Refactor: 카테고리 컴포넌트
* Categories -> CategoryBar로 컴포넌트명 변경

* CategoryBar의 스와이퍼 영역 속성 값 변경

* 현재 카테고리에 따른 스타일 변경 적용 안되는 이슈 수정
