+++ src/app/globals.css
@@ -65,3 +65,7 @@ body,
     opacity: 1;
   }
 }
+
+header .swiper-slide {
+  width: max-content;
+}
